requirements.txt
Flask
Flask-Cors
openai
google-generativeai
httpx
pandas
numpy
scikit-learn
scipy
python-dateutil
pytz
PyYAML
requests
python-dotenv
notion-client==1.0.0
PyJWT
bcrypt
chromadb
celery
redis
flower
psycopg2-binary
SQLAlchemy
alembic
# RAG System Dependencies
langchain
unstructured
pypdf
python-docx
markdown
sentence-transformers
